test: move stock price schema coverage to integration (#356)
